# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * SpaceX IPO Access API
 * OpenAPI spec version: 0.1.0
 */
export type HealthStatusStatus = typeof HealthStatusStatus[keyof typeof HealthStatusStatus];


export const HealthStatusStatus = {
  ok: 'ok',
} as const;

export interface HealthStatus {
  status: HealthStatusStatus;
}

export interface ErrorResponse {
  error: string;
}

export interface InvestorInput {
  /** @minLength 1 */
  fullName: string;
  email: string;
}

export type InvestorStatus = typeof InvestorStatus[keyof typeof InvestorStatus];


export const InvestorStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Investor {
  id: number;
  fullName: string;
  email: string;
  status: InvestorStatus;
  createdAt: string;
}

export interface InvestorCount {
  count: number;
}

export type InvestorAdminStatus = typeof InvestorAdminStatus[keyof typeof InvestorAdminStatus];


export const InvestorAdminStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface InvestorAdmin {
  id: number;
  fullName: string;
  email: string;
  status: InvestorAdminStatus;
  createdAt: string;
  shares: string;
  avgCost: string;
}

export type InvestorStatusInputStatus = typeof InvestorStatusInputStatus[keyof typeof InvestorStatusInputStatus];


export const InvestorStatusInputStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface InvestorStatusInput {
  status: InvestorStatusInputStatus;
}

export interface CreditInput {
  /** @minimum 0 */
  shares: number;
  /** @minimum 0 */
  pricePerShare: number;
}

export interface SignInInput {
  email: string;
}

export type SignInResultStatus = typeof SignInResultStatus[keyof typeof SignInResultStatus];


export const SignInResultStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface SignInResult {
  status: SignInResultStatus;
  fullName: string;
  email: string;
}

export type GetStockHistoryPeriod = typeof GetStockHistoryPeriod[keyof typeof GetStockHistoryPeriod];


export const GetStockHistoryPeriod = {
  '1D': '1D',
  '1W': '1W',
  '1M': '1M',
  '3M': '3M',
  '1Y': '1Y',
  '5Y': '5Y',
} as const;

export interface StockQuote {
  symbol: string;
  price: number;
  change: number;
  changePct: number;
  volume: number;
  marketCap: number;
  bid: number;
  ask: number;
  dayLow: number;
  dayHigh: number;
  week52High: number;
  week52Low: number;
  peRatio: number;
  avgVolume: number;
  sharesOut: string;
}

export interface StockDataPoint {
  time: string;
  price: number;
}

export interface StockHistory {
  period: string;
  data: StockDataPoint[];
}

export interface Holdings {
  investorId: number;
  shares: string;
  avgCost: string;
  updatedAt: string;
}

export type CreateDepositInputCoin = typeof CreateDepositInputCoin[keyof typeof CreateDepositInputCoin];


export const CreateDepositInputCoin = {
  BTC: 'BTC',
  ETH: 'ETH',
  DOGE: 'DOGE',
} as const;

export type DepositRecordMethod = typeof DepositRecordMethod[keyof typeof DepositRecordMethod];


export const DepositRecordMethod = {
  card: 'card',
  crypto: 'crypto',
} as const;

export interface CreateDepositInput {
  email: string;
  /** @minimum 1 */
  amount: number;
  method: DepositRecordMethod;
  coin?: CreateDepositInputCoin;
}

export type DepositRecordStatus = typeof DepositRecordStatus[keyof typeof DepositRecordStatus];


export const DepositRecordStatus = {
  pending: 'pending',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface DepositRecord {
  id: number;
  investorId: number;
  email: string;
  amount: string;
  method: DepositRecordMethod;
  coin?: string | null;
  status: DepositRecordStatus;
  createdAt: string;
}

export interface DepositAdmin {
  id: number;
  investorId: number;
  fullName: string;
  email: string;
  amount: string;
  method: DepositRecordMethod;
  coin?: string | null;
  status: DepositRecordStatus;
  createdAt: string;
}

export interface DepositStatusInput {
  status: DepositRecordStatus;
}

export interface DepositAddress {
  coin: string;
  address: string;
  updatedAt: string;
}

export interface UpdateDepositAddressInput {
  /** @minLength 10 */
  address: string;
}

export type GetStockHistoryParams = {
period?: GetStockHistoryPeriod;
};

export type GetHoldingsParams = {
email: string;
};