# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SpaceX IPO Access API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/investors/{id}/status:
    patch:
      operationId: adminUpdateInvestorStatus
      tags: [admin]
      summary: Approve or reject an investor (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestorStatusInput"
      responses:
        "200":
          description: Updated investor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvestorAdmin"
        "404":
          description: Investor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/investors/{id}/credit:
    post:
      operationId: creditInvestor
      tags: [admin]
      summary: Manually credit shares to an investor (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreditInput"
      responses:
        "200":
          description: Updated holdings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Holdings"
        "404":
          description: Investor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/deposits/{id}/status:
    patch:
      operationId: adminUpdateDepositStatus
      tags: [admin]
      summary: Verify or reject a deposit (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositStatusInput"
      responses:
        "200":
          description: Updated deposit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositRecord"
        "404":
          description: Deposit not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/deposit-addresses/{coin}:
    put:
      operationId: updateDepositAddress
      tags: [admin]
      summary: Update a crypto deposit address (admin)
      parameters:
        - name: coin
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDepositAddressInput"
      responses:
        "200":
          description: Updated address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositAddress"
        "400":
          description: Invalid coin or address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    InvestorAdmin:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        status:
          $ref: "#/components/schemas/InvestorAdminStatus"
        createdAt:
          type: string
          format: date-time
        shares:
          type: string
        avgCost:
          type: string
      required: [id, fullName, email, status, createdAt, shares, avgCost]
    InvestorAdminStatus:
      type: string
      enum: [pending, approved, rejected]
    InvestorStatusInput:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/InvestorStatusInputStatus"
      required: [status]
    InvestorStatusInputStatus:
      type: string
      enum: [pending, approved, rejected]
    CreditInput:
      type: object
      properties:
        shares:
          type: number
          minimum: 0
        pricePerShare:
          type: number
          minimum: 0
      required: [shares, pricePerShare]
...
    Holdings:
      type: object
      properties:
        investorId:
          type: integer
        shares:
          type: string
        avgCost:
          type: string
        updatedAt:
          type: string
          format: date-time
      required: [investorId, shares, avgCost, updatedAt]
...
    DepositRecordMethod:
      type: string
      enum: [card, crypto]
    DepositRecordStatus:
      type: string
      enum: [pending, completed, failed]
    DepositRecord:
      type: object
      properties:
        id:
          type: integer
        investorId:
          type: integer
        email:
          type: string
        amount:
          type: string
        method:
          $ref: "#/components/schemas/DepositRecordMethod"
        coin:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/DepositRecordStatus"
        createdAt:
          type: string
          format: date-time
      required: [id, investorId, email, amount, method, status, createdAt]
...
    DepositStatusInput:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/DepositRecordStatus"
      required: [status]
    DepositAddress:
      type: object
      properties:
        coin:
          type: string
        address:
          type: string
        updatedAt:
          type: string
          format: date-time
      required: [coin, address, updatedAt]
    UpdateDepositAddressInput:
      type: object
      properties:
        address:
          type: string
          minLength: 10
      required: [address]