# lib/api-zod/src/generated/types/creditInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * SpaceX IPO Access API
 * OpenAPI spec version: 0.1.0
 */

export interface CreditInput {
  /** @minimum 0 */
  shares: number;
  /** @minimum 0 */
  pricePerShare: number;
}