# lib/api-zod/src/generated/types/investorStatusInputStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * SpaceX IPO Access API
 * OpenAPI spec version: 0.1.0
 */

export type InvestorStatusInputStatus = typeof InvestorStatusInputStatus[keyof typeof InvestorStatusInputStatus];


export const InvestorStatusInputStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;